feat: added android auto base screen
